Change test_transformer_engine.py uses pytest.mark.mpi.
So it can be picked up by CI, similar to test_multidevice.py.
